add validation to SelectableMixin
